add additional inheritance level for json output
