+++ lib/wrf_library/json_converter.rb
@@ -1,5 +1,6 @@
 module WrfLibrary
   require "json"
   require "wrf_library/json_converter/base_station_converter"
+  require "wrf_library/json_converter/wrf_station_converter"
   require "wrf_library/json_converter/wrf_converter"
 end
